moved configuration logic to a separate classes
