+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ExploreV2/Scripts/Sections/PlacesAndEventsSection/SubSections/HighlightsSubSection/HighlightsSubSectionMenu/HighlightsSubSectionComponentController.cs
@@ -174,7 +174,7 @@ public class HighlightsSubSectionComponentController : IHighlightsSubSectionComp
 
     internal void JumpInToPlace(HotSceneInfo placeFromAPI)
     {
-        ExplorePlacesUtils.JumpInToPlace(placeFromAPI);
+        PlacesCardsConfigurator.JumpInToPlace(placeFromAPI);
         view.HidePlaceModal();
 
         dataStore.exploreV2.currentVisibleModal.Set(ExploreV2CurrentModal.None);
@@ -184,7 +184,7 @@ public class HighlightsSubSectionComponentController : IHighlightsSubSectionComp
 
     internal void JumpInToEvent(EventFromAPIModel eventFromAPI)
     {
-        ExploreEventsUtils.JumpInToEvent(eventFromAPI);
+        EventsCardsConfigurator.JumpInToEvent(eventFromAPI);
         view.HideEventModal();
 
         dataStore.exploreV2.currentVisibleModal.Set(ExploreV2CurrentModal.None);
