+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ExploreV2/Scripts/Sections/PlacesAndEventsSection/SubSections/PlacesSubSection/PlacesSubSectionMenu/PlacesSubSectionComponentController.cs
@@ -132,7 +132,7 @@ public class PlacesSubSectionComponentController : IPlacesSubSectionComponentCon
 
     internal void JumpInToPlace(HotSceneInfo placeFromAPI)
     {
-        ExplorePlacesUtils.JumpInToPlace(placeFromAPI);
+        PlacesCardsConfigurator.JumpInToPlace(placeFromAPI);
         view.HidePlaceModal();
 
         dataStore.exploreV2.currentVisibleModal.Set(ExploreV2CurrentModal.None);
